requirements.txt
pdfplumber
flask
gunicorn
python-dotenv
langchain
langchain-mistralai
langchain-openai
langchain-community
sqlalchemy
oracledb
yfinance